Store ID as static variable and increment
There have been reports such as https://wordpress.org/support/topic/doesnt-load-uncaught-typeerror/ where the IDs have fallen out of sync. This probably can result if the site uses caching in a way in which different parts of the page are cached.

As a workaround, using a static variable means that - most likely - instance IDs are to remain  the same (and will always be `1` if only one posterboard is displayed).